Major cleaning up of style as dict, attributes, element behavior.
makeStyle function moved global.
Style has only influence on the appearance of TextBox, no longer on the
enclosed FormattedString textbox.fs. It is up to the caller to make
sure that the string has the right format(s).
Adding lots of comments.
Testing on FormattedString leading and overflow.
Added some temp testing apps and .md sources.