Major cleaning up of style as dict, attributes, element behavior.
makeStyle function moved global.
Style has only influence on the appearance of TextBox, no longer on the
enclosed FormattedString textbox.fs. It is up to the caller to make
sure that the string has the right format(s).
Adding lots of comments.
Testing on FormattedString leading and overflow.
Added some temp testing apps and .md sources.


Former-commit-id: 67fd8b78b302c186a8c1784004f3ce47841a5a96
Former-commit-id: f04bf72c0738c7514a2e28dbfd999d7dc5e891d2
Former-commit-id: 8699fd0a11a18e903803de18e52c9d566da4a88a