Add separate document.docW
Add optional drawing of page frame and crop marks.
Origin of drawing elements is bottom-left (reverse of top-left could be
future option, but not now).
DrawBot new page is called by document in its own size. In case the
page has the same size, not cropmarks are drawn. In case the size of
the document is larger, then we want all PDF pages to be the same size,
and allow the page instances to draw their crop-marks and additional
information.
Page is drawing in offset, in case document.docW != page.w and
document.docH != page.h
Make grid draw rightly is page is in offset.
Add comments.


Former-commit-id: a257db119ae00baa08b98a49b9751645c3eae114
Former-commit-id: 6a961b861d43ed3adab35c59135f12f5f7328a49
Former-commit-id: d62b040e21a974208ae9c801a6997ff3c6d8a9d0