Add separate document.docW
Add optional drawing of page frame and crop marks.
Origin of drawing elements is bottom-left (reverse of top-left could be
future option, but not now).
DrawBot new page is called by document in its own size. In case the
page has the same size, not cropmarks are drawn. In case the size of
the document is larger, then we want all PDF pages to be the same size,
and allow the page instances to draw their crop-marks and additional
information.
Page is drawing in offset, in case document.docW != page.w and
document.docH != page.h
Make grid draw rightly is page is in offset.
Add comments.