Add scaleX and scaleY attribute to Page.style class. Show Healthy page4 as thumbnail on cover page1
This allows pages to be drawn as scaled thumbnails on other pages.
Needs some adjustments, where scale (or even transformation matrix)
should be part of placement, not part of element style.
The element should never be aware in what scale it is drawn.
TODO: there is some overlap with the scaling principle of image drawing.