Add scaleX and scaleY attribute to Page.style class. Show Healthy page4 as thumbnail on cover page1
This allows pages to be drawn as scaled thumbnails on other pages.
Needs some adjustments, where scale (or even transformation matrix)
should be part of placement, not part of element style.
The element should never be aware in what scale it is drawn.
TODO: there is some overlap with the scaling principle of image drawing.

Former-commit-id: b752223930fd0e3a7728c4b15e7d6d2769c68ff5 [formerly e9a1d1a68494c62b0693bc42ac32515d9d940002]
Former-commit-id: 02825be1b8aeb6d0c9aca956f717b783c9913c33
Former-commit-id: f1526223df2cdde51eac6801532e01a86849e1bb
Former-commit-id: 8b10cde009d797116ce79e56b921a92193b541cf