Major rebuild: all positions become attributes of elements, not their parent containers.
The minor disadvantage is that elements can only be placed once.
But the big advantage is that the ancestors tree is available and it is
much easier to do position evaluate through conditions. And alter the
layout based on evaluation values.
Needs many adjustments still to get all previous pages to work.


Former-commit-id: 31188f75df4396104208ea3f57f21fd9532654cf [formerly 252bee14a81b4d338113eefa143e9aebdf2dc71b]
Former-commit-id: 32175d9a308084bdc0172b82f1174ed6c05cdb02
Former-commit-id: 223caf310e46fc622d83a382d4c290aa677703c2
Former-commit-id: 44dfc50337ef09b2711fc2abda5f41346c4fab34