Major rebuild: all positions become attributes of elements, not their parent containers.
The minor disadvantage is that elements can only be placed once.
But the big advantage is that the ancestors tree is available and it is
much easier to do position evaluate through conditions. And alter the
layout based on evaluation values.
Needs many adjustments still to get all previous pages to work.