Standardizing the use of newString, based on context of the caller.
newString is now in BaseContext, with generalised API.
Removed BabelString.html references. Now HtmlString is a full
BabelString subclass. Usage of string should make the right type from
start. It makes not sense doing BabelString conversion, as they
supposed to come from the same source.
Add STRING_CLASS and NEW_STRING to contexts, so the
BaseContext.newString knows which one to take, depending on the type of
context class.
BabelString hold a ref to context now, not builder.


Former-commit-id: f35819d2a1834689a6668e1756af010b40dcab7f [formerly 59e644e1a7e72b984b10e2ae83ab27fc969ecd22]
Former-commit-id: 1762a7886305d8e0ce526d44778cc6fa332c0e6f
Former-commit-id: 995fe5c71d957ceaed0c8cc70b003ebd403a3418
Former-commit-id: 57102acc14a630cce9b5987e893deab32d8a7857