Standardizing the use of newString, based on context of the caller.
newString is now in BaseContext, with generalised API.
Removed BabelString.html references. Now HtmlString is a full
BabelString subclass. Usage of string should make the right type from
start. It makes not sense doing BabelString conversion, as they
supposed to come from the same source.
Add STRING_CLASS and NEW_STRING to contexts, so the
BaseContext.newString knows which one to take, depending on the type of
context class.
BabelString hold a ref to context now, not builder.