Debugging FlatContext by adding docTests
Problem is that the color-file format match is so strict, that it is
not possible to make drawing that output in various formats. That is
different from the original DrawBot/PageBot setup, where the drawing
can be made, without know the output format in advance.
Question is if we are going to solve this now, or we investigate
another non-DrawBot context to see if it works better there.


Former-commit-id: 3193a73af6c84b9c546391425a93abaaa8f4fc79 [formerly 48beb062043433ff40d266839036f8829db75bc8]
Former-commit-id: c40bec9577cf3b5b33e72ac255145d9f34891be6
Former-commit-id: a8e2a94f1ab97ea00e0b4bd4ac8d9b690a432498
Former-commit-id: 8bbc2c978ece5dcc1dfe50a307028ff47ccf2fdc