Debugging FlatContext by adding docTests
Problem is that the color-file format match is so strict, that it is
not possible to make drawing that output in various formats. That is
different from the original DrawBot/PageBot setup, where the drawing
can be made, without know the output format in advance.
Question is if we are going to solve this now, or we investigate
another non-DrawBot context to see if it works better there.