Make NoneDrawBotBuilder class to test DrawBot docTests if DrawBot is not available on a platform
Make pbruler docTest work in Drawbot as well as Flat context.
(context.newPage( ) needs to be called separately in the docTest, as
normall the view.build will do that).
The context view document and page are created inside the
build-tree-call squence, telling the context to make document and pages
when exporting the document. See them as virtual canvasses. In
DrawBotCanvas and FlatCanvas these can be pixel files, in the
HTMLContext this is a text output file.