Make NoneDrawBotBuilder class to test DrawBot docTests if DrawBot is not available on a platform
Make pbruler docTest work in Drawbot as well as Flat context.
(context.newPage( ) needs to be called separately in the docTest, as
normall the view.build will do that).
The context view document and page are created inside the
build-tree-call squence, telling the context to make document and pages
when exporting the document. See them as virtual canvasses. In
DrawBotCanvas and FlatCanvas these can be pixel files, in the
HTMLContext this is a text output file.


Former-commit-id: ddbe30f970fe9ff3607277410c421d41ddb9146b [formerly 43fa17e37e548de7664f6f51aabe54b501c0be8b]
Former-commit-id: a3daf7dc5fb5a2295e44a86c06cb6d31b1677150
Former-commit-id: d4cf434dc9a57953e27c099290b3da0bd80179d8
Former-commit-id: 3c197c616cb03f3be4d6a1e2c4a6086006c99b19