Debugging text positions in textBox
Solving bug that did not calculate the bounding box right for text
where tracking != 0. Now using the pixelBounds that is more accurate.
Tracking in combination with context.textSize(bs) seems to as white
space to the right.
Solved wrong calling of Element.clearElements() so in the Image
element, self._elements was not initialised.
Add pixelBounds(s, w, h) as global function in DrawBotString.
Add example poster template for DDS—> Robothon 2018. Adding test script
to show text position and bounds.


Former-commit-id: 6e68512dc1f93617feae5ae9ac99533b079a316a [formerly 0aeb48b6a0670884948f922106f6fb7ffbee37a6]
Former-commit-id: 5f86c73314f7ef5fe37c2cfdc6c30bd21bb56eee
Former-commit-id: 80be416900fa3c6db23a39c6e684a0858e950a28
Former-commit-id: bf3e2c84171814e34574f788736df845ddedc4bc