Debugging text positions in textBox
Solving bug that did not calculate the bounding box right for text
where tracking != 0. Now using the pixelBounds that is more accurate.
Tracking in combination with context.textSize(bs) seems to as white
space to the right.
Solved wrong calling of Element.clearElements() so in the Image
element, self._elements was not initialised.
Add pixelBounds(s, w, h) as global function in DrawBotString.
Add example poster template for DDS—> Robothon 2018. Adding test script
to show text position and bounds.