Major update, solving the problem of finding resource of different platforms
Test fonts and text image when in side the Lib, so the can be found
independent of installation was done by Lib copy, setup.py to egg or
linked by .pth file in site-packages.
The nature of the resources is that they remain small, compared to the
growing Examples folder (which is now removed to the PageBotExamples
repository.
contexts/platform now contains various functions to get the PageBot/Lib
getRootPath( ), getTestFontsPath( ), getResourcesPath( ), etc.
Font collection is now cached and based on the current platform.
Many docTests added.
Cleaned DrawBotContext and FlatContext and making API more consistent.
Split fitting function DrawBotContext newString to separate width and
height seeking class methods.
Many doc[0] page references changed to doc[1], as now the first page is
default 1 (as right page) instead of the default list index 0 (even
page number should not be a right page).
Cleaning of obsolete and deprecated files.


Former-commit-id: b4d08099fbec47f18f1592e72cd94795e7374712 [formerly 0f5a86393b5bae02a82a142ad37267ee53c1de0e]
Former-commit-id: cb486112657e2e3faa930ef933d5e6b312b72fa1
Former-commit-id: 6eba8e4cd97afc617abfb06220e9620f75add86f
Former-commit-id: 78f12f5302cf5f3c971fc4edd1c1a496ebb2872e