Major update, solving the problem of finding resource of different platforms
Test fonts and text image when in side the Lib, so the can be found
independent of installation was done by Lib copy, setup.py to egg or
linked by .pth file in site-packages.
The nature of the resources is that they remain small, compared to the
growing Examples folder (which is now removed to the PageBotExamples
repository.
contexts/platform now contains various functions to get the PageBot/Lib
getRootPath( ), getTestFontsPath( ), getResourcesPath( ), etc.
Font collection is now cached and based on the current platform.
Many docTests added.
Cleaned DrawBotContext and FlatContext and making API more consistent.
Split fitting function DrawBotContext newString to separate width and
height seeking class methods.
Many doc[0] page references changed to doc[1], as now the first page is
default 1 (as right page) instead of the default list index 0 (even
page number should not be a right page).
Cleaning of obsolete and deprecated files.