Cleaning HtmlBuilder and example web elements
Context is now stored in the current view. This gives better control on
the type of context related to a type of view.
Also it allows to have different contexts inside views that are placed
on a page.
Consequence is that getContext() in only need from now for example
scripts that don't include a Document to be created.
Make exporting a website to local folder work.
Add support for HtmlBuilder <footer>
Clear further for cssClass. Add cssId for separate control on Html
#name output.
Add cssId attribute in Element for HtmlBuilder, as separate #id tag
attributes.


Former-commit-id: 322f804e03fd101d6b42627e0fb7fc40c8483928 [formerly c54dd784e5eed1e7c3258fecf705615d55daa3d4]
Former-commit-id: 52f513027527fbb65f9b9eb211eeae612a61f99a
Former-commit-id: 1cd34a8f632002b2b50dbc493098c2facdcdde36
Former-commit-id: 694fa1fb3f1563596196dff475a919eb76d16224