Cleaning HtmlBuilder and example web elements
Context is now stored in the current view. This gives better control on
the type of context related to a type of view.
Also it allows to have different contexts inside views that are placed
on a page.
Consequence is that getContext() in only need from now for example
scripts that don't include a Document to be created.
Make exporting a website to local folder work.
Add support for HtmlBuilder <footer>
Clear further for cssClass. Add cssId for separate control on Html
#name output.
Add cssId attribute in Element for HtmlBuilder, as separate #id tag
attributes.