Debugging doctests
Add functionality to unit.py uRound, so it takes lists and arguments
too, answering the same structure of the original tuples. All numbers
convert to pt-units and all units are rounded.
Implement the special version of self.w and self.h in some Element
classes.
Add units to most doctests metrics, as far as they are not
auto-converted to pt-units.