Debugging doctests
Add functionality to unit.py uRound, so it takes lists and arguments
too, answering the same structure of the original tuples. All numbers
convert to pt-units and all units are rounded.
Implement the special version of self.w and self.h in some Element
classes.
Add units to most doctests metrics, as far as they are not
auto-converted to pt-units.


Former-commit-id: 1c3d3a63499dd989601c7242a050b94321007df4 [formerly 776ba302ea17aacac58f0432ee8dffb92285e497]
Former-commit-id: 7f37a904fcc516f1e7d1babfd085a3bbf016915d
Former-commit-id: 96d56023aa6f2e81774f5723bd1b12bb572870ce
Former-commit-id: f396105ab9e8782f6811ea7fb4668f7cf0ff8a2d