Remove min/max clipping for position and size.
Too complex and too easy to make errors as there is deep-down clipping
on points and sizes going on.
Also there was a problem when subtracting from a size, where the unit
temporarily gets negative and is then clipped in mid-calculation.
Removing all (automatic) clipping on position and size makes the model
easier to understand.
Another problem was that it's not really consistent which min/max
should be passed on with unit arithmetic.
It also reduces the amount of parameters for units and elements.
Overall system limit is not simply -XXXL and XXXL


Former-commit-id: c97d69f1ee5326522e7f87dfb3ba685d997822f5 [formerly 4833c90cf92e12f67c44aa32ff6a25f267904a76]
Former-commit-id: b10f9a4ea9daa91bdc7a6bee006b244f742ee1e0
Former-commit-id: d9177c37196bb19f2a8a71370a06af4bbc6e8cd3
Former-commit-id: 8b5f237382a41241abc70537090720a986866483