Remove min/max clipping for position and size.
Too complex and too easy to make errors as there is deep-down clipping
on points and sizes going on.
Also there was a problem when subtracting from a size, where the unit
temporarily gets negative and is then clipped in mid-calculation.
Removing all (automatic) clipping on position and size makes the model
easier to understand.
Another problem was that it's not really consistent which min/max
should be passed on with unit arithmetic.
It also reduces the amount of parameters for units and elements.
Overall system limit is not simply -XXXL and XXXL