Solve for Travis #673
For now, comment the path calls in Font and Glyph that need a context.
As we don't want the context to be stored in Font or Glyph, it will be
an attribute for Glyph.getBezierPath(context),
Glyph.getFlattenedPath(context) and
Glyph.getFlattenedContours(context). For that the Glyph._initialize
needs to be changed, as at the moment of object construction the
context is not known.
For now this should fix the Travis errors.


Former-commit-id: 63de520b2fc223e813154bed5f15685343101d01 [formerly 472708fd6190baa47eb1d796d44af4f4c40a81c0]
Former-commit-id: 95e22488cccccea7a9756dab2ec655a472f5978b
Former-commit-id: c6530585b3c1458288f31e466c3054a31bf990a4
Former-commit-id: 7bd3c65f7baa3b18d634db282615b47f32176f20