Solve for Travis #673
For now, comment the path calls in Font and Glyph that need a context.
As we don't want the context to be stored in Font or Glyph, it will be
an attribute for Glyph.getBezierPath(context),
Glyph.getFlattenedPath(context) and
Glyph.getFlattenedContours(context). For that the Glyph._initialize
needs to be changed, as at the moment of object construction the
context is not known.
For now this should fix the Travis errors.