Major update
All view flags are now element properties. This gives the option to
apply view flags to a single page or elements, while the view (which is
also an element) can hold the global setting of the viewing properties.
Extended to GRID and BASE options, so it also indicates if lines,
indices or measures of grid and baseline should be drawn on foreground
or background. This is a flexible method to get part of the grid drawn
on a foreground and part on background.
This approach also makes it easier to use views as illustration on a
page (e.g. when making manuals).
Many of the view options are renamed to become more general usable.
E.g. view.drawPageFrame and e.drawElementFrame now collapsed into
e.drawFrame.
All existing examples in PageBot repository have been renamed
accordingly.