Major update
All view flags are now element properties. This gives the option to
apply view flags to a single page or elements, while the view (which is
also an element) can hold the global setting of the viewing properties.
Extended to GRID and BASE options, so it also indicates if lines,
indices or measures of grid and baseline should be drawn on foreground
or background. This is a flexible method to get part of the grid drawn
on a foreground and part on background.
This approach also makes it easier to use views as illustration on a
page (e.g. when making manuals).
Many of the view options are renamed to become more general usable.
E.g. view.drawPageFrame and e.drawElementFrame now collapsed into
e.drawFrame.
All existing examples in PageBot repository have been renamed
accordingly.


Former-commit-id: 9705602490e4972145fc6699df2e6752f6b63161 [formerly 06452def258dfa5da79704c621d0ab67301d4a5e]
Former-commit-id: e741e03abba5f93507b9ab2ab96eb70569a69098
Former-commit-id: 0169bca4d87ff143a48e6f5258ec584a17b77aea
Former-commit-id: fd81a1b7277d5c2605864f65dafd778ccc5c7ec2