Cleaning examples to show options for GRID and BASE display.
Some view showings are called twice (background and foreground) to
allow GRID and BASE parts to be drawn independently on foreground and
background (e.g. BASE_LINE or BASE_LINE_BG as part of the
e.showBaselines set).
Add e.language and e.hyphenation as properties on the element style.