Cleaning examples to show options for GRID and BASE display.
Some view showings are called twice (background and foreground) to
allow GRID and BASE parts to be drawn independently on foreground and
background (e.g. BASE_LINE or BASE_LINE_BG as part of the
e.showBaselines set).
Add e.language and e.hyphenation as properties on the element style.


Former-commit-id: 7aafe3b1c63d256b16b5c9caface7df6b191fc26 [formerly 07ad345d82699913984ce496417ae72c82ada4c6]
Former-commit-id: b35c7147805bb26edbdb05ce6ac3340320410a02
Former-commit-id: c99ffd2bd82d7e038471e29fe8992957da93ecbd
Former-commit-id: 0c8737ff82ca32074c5fcc24ea99b443b307ab40