TextBox: solved bug in position if baseline index markers
Remove option to show both index and baseline position on the same side
of a column.
Add 2 examples, showing the rotation of text boxes, including the
rotated drawing of baseline and baseline markers.
Add remark TODO for BinaryData to create the _export directory if it
does not exist. This is caused by not using a Document.export( ) which
normally create the _export folder when building.