TextBox: solved bug in position if baseline index markers
Remove option to show both index and baseline position on the same side
of a column.
Add 2 examples, showing the rotation of text boxes, including the
rotated drawing of baseline and baseline markers.
Add remark TODO for BinaryData to create the _export directory if it
does not exist. This is caused by not using a Document.export( ) which
normally create the _export folder when building.


Former-commit-id: 69e08b607f5ad622d8c16e523996de32b7ce059e [formerly 0bc676268743c4a34b754411f31dae3250b145e3]
Former-commit-id: 4daad027b4cd50afd97005076e3c9130b8339684
Former-commit-id: f2f00987cc647ec2e8dc18ff2b109d0212b689a8
Former-commit-id: 4053cac602221df3d0d86c672d67b0482e7ac134