PageView: Add drawing of 1 type of color bar on left side of page.
The usage of these calibration color bars in PageBot is limited, for
now, as most offset printing companies want to add their own, specific
for their printing presses. However it is good to be able to show them
for educational purposes and e.g. as illustrations in PageBot generated
manuals.
Move some parameters in pagebot.contants up, so they can be used by
other parts of the file.
Add 100_PageColorBars.py as example script for empty pages  with grid
and color calibration on one side.