PageView: Add drawing of 1 type of color bar on left side of page.
The usage of these calibration color bars in PageBot is limited, for
now, as most offset printing companies want to add their own, specific
for their printing presses. However it is good to be able to show them
for educational purposes and e.g. as illustrations in PageBot generated
manuals.
Move some parameters in pagebot.contants up, so they can be used by
other parts of the file.
Add 100_PageColorBars.py as example script for empty pages  with grid
and color calibration on one side.


Former-commit-id: 13a179d40683b7eddef72fc71146f42813a89d4c [formerly 2f8196c1d6b5b6f226b13c30bd4eaf94fa1b8edb]
Former-commit-id: f3fb75753cf824f9005fce90846527aa8b775186
Former-commit-id: 412c4cf79d642f7ea8ec9aca173e03a3299cf367
Former-commit-id: da401bdd01f488bfbc05ef1b117610e019f326ea