Add automatic scaling and saving of cached images in _scaled folder.
The DrawBot Context now knows how to create a scaled image path name:
DrawBotContext.path2ScaledImagePath(path, w, h)
The Image class is saving to a cached file, if the flag is (default) on
and if the scaling is over a 0.8 — 1.2 range in either direction.
Add Element.prepare(view) recursive broadcast to all child elements of
a document and page, to prepare for building. This Image is using this
call to make sure all images exist in the caching size that is needed.
The initial call for the page.prepare(view) is done by the PageView,
jut before any page.build is called.
Add _local to .gitignore
Add more doctests


Former-commit-id: fdc20b61d3da8c6977b529c297cb238f3e2e0f22 [formerly 4b80950c06965fda94b76392abcabb6f03c7409b]
Former-commit-id: 8c85ef55209f5f42792e9b8d1a695018bebf5b81
Former-commit-id: 8cab34767a45a4ff681316147e63dd21f95da7d2
Former-commit-id: b0e9898cdc59b8e7ca7870999d19a4e20e2a19ab