Solving many old bug related to originTop parameter.
Add _images to .gitignore
Working on SketchContext, to read Sketch documents.
Solved bug that did not pass the document.originTop parameter to new
pages.
Update many element tests related to originTop parameter.
The Document.originTop, page.originTop and Element.originTop are now
write-only properties. They can be set during creation of the instance,
but are not supposed to be changed later.
Solved problem for PageView not showing get padding right for
originTop=True.
originTop is now stored in the Document and Element as _originTop. The
parameter is removed from the style and root style.


Former-commit-id: 5e0818ae041727d61c2d7fac4c3a525b6b5583ce [formerly 1221d6c222732fc60d2bb55b463ad38dbd1cea28]
Former-commit-id: 6b618f6006fdc3d69d30afba77a4bb77462b3863
Former-commit-id: 4d3bb2cc271eadf4fbe7952329ef90798a45ac4d
Former-commit-id: f768414210df4cfc0a0f7c6383785985e90c2277