Add color.isRgba and color.rgba property
This ways color.rgb and color.rgba can be compared. “a” is valid if
opacity < 1.
Add Element.proportional property, stored in the style. It sets the
Element.w, Element.h and Element.d proportionally if any of them is
changed. Same happens with the scales.
Note that since the Element.proportional does not work on margins and
puddings, those measures won’t stay proportional. This just applies to
plain size and scale.
Add Nanosite element Collection
Solved old bug that did not show transparency on string drawing in
DrawBot.
Add style attribute to HtmlBuilder of tags. TODO: make it optional to
supply a dictionary here that will be translated into HTML-tag-style
syntax.


Former-commit-id: f35ca9d8492e2854033596423afc8bb6231eb2e3 [formerly f6b62cf0a9294b9388089431c7c468653e42b796]
Former-commit-id: b81c8170b475cf849fca9ac15f9ccb07c0b8f9a0
Former-commit-id: ce2f9e9b0bb23d1ba379bafb9e4aacc46a50506a
Former-commit-id: 7a7ca031d4079a436b0b75dd6780362edbcc207b