Adding Typesetter.tagStack that holds names of the current open tag names
Rename style attribute firstParagraphIndent to firstTagIndent to define
the indent of a paragraph if its name is different from the previous
tag. This allows to not indent after headlines, overwriting the
firstLineIndent value.
Add default BabelString.columnStart(firstColumnIndent) to answer
unchanged self.
Probably solved the indenting issues for MarkDown in body text
situations.


Former-commit-id: da1295bcc8d5ccdab9e6f35efd44a47046348f26 [formerly a75548c05d199e4a0df2310c0ecef8dbe5148f50]
Former-commit-id: 0ae88ad6e028ccd53348e22ddf76b8077229dc9c
Former-commit-id: 6cc0b4447d6776f7933d7e1a9d867ee2e1b17b61
Former-commit-id: 42345cd989efc4e05cd61f1f033f8fe41795bb3f