Adding some view attributes as Element property, so they can be locally overwritten
This is forcing the showing (and stroke color and stroke width) of
element view attribute, independent of overall view settings.
This is useful, e.g. when showing cropmarks on a DocWrap elementen,
while the main document may or may not show them.
Some cleaning done on PageView attribute showing.
Add viewFrameStroke and viewFrameStrokeWidth to style
The frame viewing parameters are independent from the border drawing of
an Element instance.


Former-commit-id: 1ee800aa2590995319ab9efce5e4b9004a985fa7 [formerly 84e99d2417b858937c2297ec0429832ba689287d]
Former-commit-id: 71cee33f085b744aceb75aaad5380d16227bbfa6
Former-commit-id: a1e3396af8a52fedd25c01bbe12ddb678ec0f9dc
Former-commit-id: 1ea32f99e35ff07de528014c9b01fca3bedf385f