Adding image scaleType parameter for better control.
Preciously the scaling type was depending on wether target w and/or h
were defined, but that often leads to confusion and ambiguous
interpretation. Furthermore, with the InDesignContent we need explicit
definition of the scaleType, regardless of (w, h).
Adding SCALE_TYPE_PROPORTIONAL and more to the constants.
Add some **kwargs to newPage and newDocument, to support more
information going into some contexts that need it (e.g. page.padding)
Replace direct imports of DrawBotContext by getContext, since we no
longer have the DrawBotContext installed in the default PageBot
Some naming of width and height into “w” and “h” to be more consistent
through the system.


Former-commit-id: 60a99c92cc00331cd20c96a64a6835d0f0c0b152 [formerly aa95aef2ea7b559e2dc6df24a0aa2816979753ce]
Former-commit-id: 4f8d54188276c52769ab5f3f223738b321294c1f
Former-commit-id: da58aa4ef2486b4d13170ef6418998b38a7696c9
Former-commit-id: eb97c8552e4c85642d85842cb323875b61e4c6e5