New versions of PageBot example TTF with typoAscender == ascender
And same for the other vertical metrics values.
This seems to solve the Sketch-->PageBot vertical metrics problem of text.
More testing to be done with other typefaces, but for now this solved the simple conversion.
Remove relative attributes in font.info again.
Rename font.py isFontPath to asFontPath, since it is doing a conversion.
SketchContext is doing the better baseline Text conversion now.
Ad optional storage of fontName in style, beside Font instance.