Replace reference to BabelText, that does not exist anymore.
The result of context.getTextLines is a list of BabelLineInfo instances.
There is also baseContext.getBaseLines, which takes the textLines and answers a dictionary of y-positions-->BabelLineInfo. SInce the position attributes are standard in a BabelLineInfo, such a method does not need to know what native line info is on board.
Still working on some more functions in FlatContext, such as textBox and overflow.