Solving problem of uninitialized context and mismatch between context type and doc.view type
Now doc.view.context initialized to the default context type of that view. And if defined, is it checked if the view <--> context combination is supported.
Add FlatContext._getValidSize to answer a validated/initialized page/document size.
Add dummy Element.prepare_flat recursive call, equivalent to the other prepare methods.